+++ src/rs/symbol.rs
@@ -18,7 +18,7 @@ pub struct Symbol {
 
     pub vram: u64,
 
-    pub size: Option<u64>,
+    pub size: u64,
 
     pub vrom: Option<u64>,
 
@@ -26,13 +26,7 @@ pub struct Symbol {
 }
 
 impl Symbol {
-    pub fn new(
-        name: String,
-        vram: u64,
-        size: Option<u64>,
-        vrom: Option<u64>,
-        align: Option<u64>,
-    ) -> Self {
+    pub fn new(name: String, vram: u64, size: u64, vrom: Option<u64>, align: Option<u64>) -> Self {
         Self {
             name,
             vram,
@@ -46,7 +40,7 @@ impl Symbol {
         Self {
             name,
             vram,
-            size: None,
+            size: 0,
             vrom: None,
             align: None,
         }
@@ -57,11 +51,7 @@ impl Symbol {
     }
 
     pub fn get_size_str(&self) -> String {
-        if let Some(size) = self.size {
-            //return format!("0x{0:X}", size);
-            return format!("{}", size);
-        }
-        "None".into()
+        format!("{}", self.size)
     }
 
     pub fn get_vrom_str(&self) -> String {
@@ -124,11 +114,11 @@ pub(crate) mod python_bindings {
     #[pymethods]
     impl super::Symbol {
         #[new]
-        #[pyo3(signature=(name,vram,size=None,vrom=None,align=None))]
+        #[pyo3(signature=(name,vram,size=0,vrom=None,align=None))]
         fn py_new(
             name: String,
             vram: u64,
-            size: Option<u64>,
+            size: u64,
             vrom: Option<u64>,
             align: Option<u64>,
         ) -> Self {
@@ -160,12 +150,12 @@ pub(crate) mod python_bindings {
         }
 
         #[getter]
-        fn get_size(&self) -> PyResult<Option<u64>> {
+        fn get_size(&self) -> PyResult<u64> {
             Ok(self.size)
         }
 
         #[setter]
-        fn set_size(&mut self, value: Option<u64>) -> PyResult<()> {
+        fn set_size(&mut self, value: u64) -> PyResult<()> {
             self.size = value;
             Ok(())
         }
@@ -212,14 +202,11 @@ pub(crate) mod python_bindings {
 
         #[pyo3(signature=(humanReadable=true))]
         fn serializeSize(&self, humanReadable: bool) -> PyResult<PyObject> {
-            Python::with_gil(|py| match self.size {
-                None => Ok(Python::None(py)),
-                Some(size) => {
-                    if humanReadable {
-                        return format!("0x{:X}", size).into_py_any(py);
-                    }
-                    size.into_py_any(py)
+            Python::with_gil(|py| {
+                if humanReadable {
+                    return format!("0x{:X}", self.size).into_py_any(py);
                 }
+                self.size.into_py_any(py)
             })
         }